Rename adult button and correct input paths
